Added more qualities
There are dependencies on race that aren't captured in the Quality
datatype. Also "Scorched" yields double BP for hackers or technomancers
but there's no way to capture that currently. Some qualities have a
specialization, like an addiction to a specific substance or being
incompetent at a particular active skill. Need to rethink the data type
to capture this sort of information.